@@ -1,11 +1,10 @@
 import 'dart:async';
-
-import 'package:arduino_ble_sensor/model/sensor_data.dart';
+import 'package:move/model/sensor_data.dart';
 import 'package:flutter/material.dart';
 
 class SensorView extends StatefulWidget {
   final SensorData sensorData;
-  SensorView(this.sensorData, {Key key}) : super(key: key);
+  SensorView(this.sensorData, {Key? key}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {
@@ -15,18 +14,18 @@ class SensorView extends StatefulWidget {
 
 class _SensorViewState extends State<SensorView> {
   String _lastTime = "now";
-  Timer _timeUpdater;
+  Timer? _timeUpdater;
 
   @override
   void initState() {
     _timeUpdater =
-        new Timer.periodic(const Duration(seconds: 1), _updateLastTime);
+        new Timer.periodic(const Duration(milliseconds: 500), _updateLastTime);
     super.initState();
   }
 
   @override
   void dispose() {
-    _timeUpdater.cancel();
+    _timeUpdater!.cancel();
     super.dispose();
   }
 
@@ -36,7 +35,11 @@ class _SensorViewState extends State<SensorView> {
       child: Column(
         children: <Widget>[
           ListTile(
-            leading: Image.asset('assets/images/temp.png'),
+            title: Text(
+          "gesture 1: left, gesture 2:jump"
+          ),
+          ),
+          ListTile(
             title: Text(
                 'result : ${widget.sensorData.result.toStringAsFixed(2)} %'),
           ),
