@@ -1,10 +1,8 @@
-import 'package:arduino_ble_sensor/sensor_list_screen.dart';
+import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
+import 'package:move/sensor_list_screen.dart';
 
-import 'bluetooth_off_screen.dart';
-//main page
-int sec = 0;
 
 void main() => runApp(MyApp());
 
@@ -12,18 +10,125 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      color: Colors.lightBlue,
-      debugShowCheckedModeBanner: false,
-      home: StreamBuilder<BluetoothState>(
-          stream: FlutterBlue.instance.state,
-          initialData: BluetoothState.unknown,
-          builder: (c, snapshot) {
-            final state = snapshot.data;
-            if (state == BluetoothState.on) {
-              return SensorListScreen();
-            }
-            return BluetoothOffScreen(state: state);
-          }),
+      title: 'MOVE!',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: MyHomePage(title: 'MOVE!'),
     );
   }
 }
+
+class MyHomePage extends StatefulWidget {
+  MyHomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+  FlutterBlue flutterBlue = FlutterBlue.instance;
+  final List<BluetoothDevice> devicesList = [];
+  final Map<Guid, List<int>> readValues = new Map<Guid, List<int>>();
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  final _writeController = TextEditingController();
+  BluetoothDevice? _connectedDevice;
+  List<BluetoothService>? _services;
+
+  _addDeviceTolist(final BluetoothDevice device) {
+    if (!widget.devicesList.contains(device)) {
+      setState(() {
+        widget.devicesList.add(device);
+      });
+    }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    widget.flutterBlue.connectedDevices
+        .asStream()
+        .listen((List<BluetoothDevice> devices) {
+      for (BluetoothDevice device in devices) {
+        _addDeviceTolist(device);
+      }
+    });
+    widget.flutterBlue.scanResults.listen((List<ScanResult> results) {
+      for (ScanResult result in results) {
+        _addDeviceTolist(result.device);
+      }
+    });
+    widget.flutterBlue.startScan();
+  }
+
+  ListView _buildListViewOfDevices() {
+    List<Container> containers = [];
+    for (BluetoothDevice device in widget.devicesList) {
+      containers.add(
+        Container(
+          height: 50,
+          child: Row(
+            children: <Widget>[
+              Expanded(
+                child: Column(
+                  children: <Widget>[
+                    Text(device.name == '' ? '(unknown device)' : device.name),
+                    Text(device.id.toString()),
+                  ],
+                ),
+              ),
+              FlatButton(
+                color: Colors.blue,
+                child: Text(
+                  'Connect',
+                  style: TextStyle(color: Colors.white),
+                ),
+                onPressed: () async {
+                  widget.flutterBlue.stopScan();
+                  try {
+                    await device.connect();
+                  } catch (e) {
+                    if (e != 'already_connected') {
+                      throw e;
+                    }
+                  } finally {
+                    _services = await device.discoverServices();
+                  }
+                  setState(() {
+                    _connectedDevice = device;
+
+                  });
+                  Navigator.push(context, MaterialPageRoute(builder: (context) => SensorListScreen()));
+                  //Navigator.push(context, MaterialPageRoute(builder: (context) => Blue_List()));
+                },
+              ),
+            ],
+          ),
+        ),
+      );
+    }
+
+    return ListView(
+      padding: const EdgeInsets.all(8),
+      children: <Widget>[
+        ...containers,
+      ],
+    );
+  }
+
+
+  ListView _buildView() {
+    return _buildListViewOfDevices();
+  }
+
+  @override
+  Widget build(BuildContext context) => Scaffold(
+    appBar: AppBar(
+        title: Text(widget.title),
+         //예에ㅔ
+          ),
+    body: _buildView(),
+  );
+}
+
